# server-go/internal/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Telemex Go API
  version: 1.0.0
  description: >
    Telemedicine backend (Go + PostgreSQL). Every response —
    success or error — uses the standard envelope
    `{success, response_code, response_description, data, message,
    request_id, path, timestamp}`. `response_code` is a numeric string:
    `00` success, `006` validation, `009` forbidden,
    `011` unauthenticated, `004` not found, `016` server error.
    See `quirks.md` for preserved frontend-compat behaviors.
servers:
  - url: http://localhost:4000/api/v1
...
security:
  - bearerAuth: []
paths:
...
  /patients/me:
    get:
      tags: [Patients]
      summary: Patient summary
      security: [bearerAuth: []]
      responses:
        '200':
          description: Summary
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/PatientSummary' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    patch:
      tags: [Patients]
      summary: Update patient profile (JSON or multipart)
      security: [bearerAuth: []]
      description: Allowlist keys only; unknown keys 400. Scalars blank leave untouched; arrays replace.
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PatientPatch' }
          multipart/form-data:
            schema:
              type: object
              properties:
                file: { type: string, format: binary }
                first_name: { type: string }
                last_name: { type: string }
                allergies:
                  { type: string, description: 'JSON array string; "" clears' }
      responses:
        '200':
          description: Updated profile
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/PatientProfile' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /patients/me/profile:
    get:
      tags: [Patients]
      summary: Full patient profile
      security: [bearerAuth: []]
      responses:
        '200':
          description: Profile
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/PatientProfile' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    patch:
      tags: [Patients]
      summary: Update patient profile (alias)
      security: [bearerAuth: []]
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PatientPatch' }
      responses:
        '200':
          description: Updated profile
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/PatientProfile' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /patients/me/timezone:
    patch:
      tags: [Patients]
      summary: Set patient timezone
      security: [bearerAuth: []]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [timezone]
              properties:
                timezone: { type: string, example: Africa/Lagos }
      responses:
        '200':
          description: Updated profile
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/PatientProfile' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /patients/me/profile-picture:
    patch:
      tags: [Patients]
      summary: Set profile picture URL
      security: [bearerAuth: []]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [profile_picture_url]
              properties:
                profile_picture_url: { type: string }
      responses:
        '200':
          description: Updated profile
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/PatientProfile' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /patients/me/profile-picture/upload:
    post:
      tags: [Patients]
      summary: Upload profile picture (5MB max)
      security: [bearerAuth: []]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file: { type: string, format: binary }
                image: { type: string, format: binary }
                profile_picture: { type: string, format: binary }
      responses:
        '200':
          description: Updated profile
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/PatientProfile' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '413': { $ref: '#/components/responses/PayloadTooLarge' }
        '415': { $ref: '#/components/responses/UnsupportedMedia' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  # --------------------------------------------------------- Doctors ----
  /doctors/me:
    get:
      tags: [Doctors]
      summary: Doctor summary
      security: [bearerAuth: []]
      responses:
        '200':
          description: Summary
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/DoctorSummary' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    patch:
      tags: [Doctors]
      summary: Update doctor profile
      security: [bearerAuth: []]
      description: 6 keys only (first_name, last_name, phone_number, profile_picture_url, license_no, specializations).
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DoctorPatch' }
      responses:
        '200':
          description: Updated profile
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/DoctorProfile' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /doctors/me/profile:
    get:
      tags: [Doctors]
      summary: Full doctor profile
      security: [bearerAuth: []]
      responses:
        '200':
          description: Profile
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/DoctorProfile' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    patch:
      tags: [Doctors]
      summary: Update doctor profile (alias)
      security: [bearerAuth: []]
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DoctorPatch' }
      responses:
        '200':
          description: Updated profile
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/DoctorProfile' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /doctors/me/timezone:
    patch:
      tags: [Doctors]
      summary: Set doctor timezone
      security: [bearerAuth: []]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [timezone]
              properties:
                timezone: { type: string, example: Africa/Lagos }
      responses:
        '200':
          description: Updated profile
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/DoctorProfile' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /doctors/me/profile-picture:
    patch:
      tags: [Doctors]
      summary: Set profile picture URL
      security: [bearerAuth: []]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [profile_picture_url]
              properties:
                profile_picture_url: { type: string }
      responses:
        '200':
          description: Updated profile
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/DoctorProfile' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /doctors/me/profile-picture/upload:
    post:
      tags: [Doctors]
      summary: Upload profile picture (5MB max)
      security: [bearerAuth: []]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file: { type: string, format: binary }
                image: { type: string, format: binary }
                profile_picture: { type: string, format: binary }
      responses:
        '200':
          description: Updated profile
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/DoctorProfile' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '413': { $ref: '#/components/responses/PayloadTooLarge' }
        '415': { $ref: '#/components/responses/UnsupportedMedia' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /doctors/patients:
    get:
      tags: [Doctors]
      summary: Doctor's patients (search + pagination)
      security: [bearerAuth: []]
      parameters:
        - $ref: '#/components/parameters/QueryQ'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Patient list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          patients:
                            type: array
                            items:
                              { $ref: '#/components/schemas/DoctorPatientRow' }
                          pagination:
                            { $ref: '#/components/schemas/Pagination' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /doctors/{id}/activate:
    patch:
      tags: [Doctors]
      summary: Activate doctor (admin)
      security: [bearerAuth: []]
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Activated profile
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/DoctorProfile' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /doctors/{id}/deactivate:
    patch:
      tags: [Doctors]
      summary: Deactivate doctor (admin, revokes sessions)
      security: [bearerAuth: []]
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Deactivated profile
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/DoctorProfile' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  # ----------------------------------------------------------- Admin ----
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    reminderKey:
      type: apiKey
      in: header
      name: x-reminder-key
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    QueryQ:
      name: q
      in: query
      schema: { type: string }
    Page:
      name: page
      in: query
      schema: { type: integer, default: 1 }
...
    Limit:
      name: limit
      in: query
      schema: { type: integer, default: 20 }
...
  responses:
    BadRequest:
      description: Validation failure (006)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Unauthorized:
      description: Missing/invalid token (011)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Forbidden:
      description: Wrong role or foreign record (009)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    NotFound:
      description: Missing record (004)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
    PayloadTooLarge:
      description: File exceeds size cap (413)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    UnsupportedMedia:
      description: Non-image content type (415)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
    ServiceUnavailable:
      description: Unconfigured provider or database (503)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  schemas:
    Envelope:
      type: object
      required:
        [
          success,
          response_code,
          response_description,
          data,
          request_id,
          path,
          timestamp,
        ]
      properties:
        success: { type: boolean, example: true }
        response_code: { type: string, example: '00' }
        response_description: { type: string, example: Success }
        data: { description: Endpoint payload (object, array, or null) }
        message:
          {
            type: array,
            items: { type: string },
            description: Validation details (400 only),
          }
        request_id: { type: string, format: uuid }
        path: { type: string }
        timestamp: { type: string, format: date-time }
    ErrorEnvelope:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - type: object
          properties:
            success: { type: boolean, example: false }
            data: { nullable: true, example: null }
...
    Pagination:
      type: object
      properties:
        total: { type: integer }
        page: { type: integer }
        perPage: { type: integer }
        total_pages: { type: integer }
        totalPages: { type: integer }
        limit: { type: integer }
...
    PatientSummary:
      type: object
      properties:
        _id: { type: string, format: uuid }
        registration_no: { type: string }
        mrn: { type: string }
        first_name: { type: string }
        last_name: { type: string }
        full_name: { type: string }
        email: { type: string }
        phone_number: { type: string }
        verified: { type: boolean }
        profile_picture_url: { type: string }
        date_of_birth: { type: string, nullable: true, example: '1990-04-01' }
        gender: { type: string, nullable: true }
        occupation: { type: string, nullable: true }
        marital_status: { type: string, nullable: true }
        timezone: { type: string, nullable: true }
        allergies: { type: array, items: { type: string } }
        previous_medical_conditions: { type: array, items: { type: string } }
    PatientProfile:
      allOf:
        - $ref: '#/components/schemas/PatientSummary'
        - type: object
          properties:
            middle_name: { type: string, nullable: true }
            address: { type: string, nullable: true }
            medical_flags: { type: array, items: { type: string } }
    PatientPatch:
      type: object
      properties:
        first_name: { type: string }
        last_name: { type: string }
        middle_name: { type: string }
        phone_number: { type: string }
        gender: { type: string }
        marital_status: { type: string }
        occupation: { type: string }
        address: { type: string }
        profile_picture_url: { type: string }
        date_of_birth: { type: string, example: '1990-04-01' }
        allergies: { type: array, items: { type: string } }
        previous_medical_conditions: { type: array, items: { type: string } }
        medical_flags: { type: array, items: { type: string } }
...
    DoctorSummary:
      type: object
      properties:
        _id: { type: string, format: uuid }
        doctor_no: { type: string }
        first_name: { type: string }
        last_name: { type: string }
        full_name: { type: string }
        email: { type: string }
        phone_number: { type: string }
        active: { type: boolean }
        specializations: { type: array, items: { type: string } }
        profile_picture_url: { type: string }
        timezone: { type: string, nullable: true }
    DoctorProfile:
      allOf:
        - $ref: '#/components/schemas/DoctorSummary'
        - type: object
          properties:
            license_no: { type: string, nullable: true }
    DoctorPatch:
      type: object
      properties:
        first_name: { type: string }
        last_name: { type: string }
        phone_number: { type: string }
        profile_picture_url: { type: string }
        license_no: { type: string }
        specializations: { type: array, items: { type: string } }
...
    DoctorPatientRow:
      type: object
      properties:
        _id: { type: string, format: uuid }
        first_name: { type: string }
        middle_name: { type: string, nullable: true }
        last_name: { type: string }
        full_name: { type: string }
        email: { type: string }
        phone_number: { type: string, nullable: true }
        registration_no: { type: string }
        mrn: { type: string }
        gender: { type: string, nullable: true }
        date_of_birth: { type: string, nullable: true }
        marital_status: { type: string, nullable: true }
        occupation: { type: string, nullable: true }
        address: { type: string, nullable: true }
        profile_picture_url: { type: string }
        verified: { type: boolean }
        consultation_id: { type: string, nullable: true }
        has_consultation_with_doctor: { type: boolean }
...